back/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-heroku
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
pipenv==2021.5.29
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
pandas==1.4.2
numpy==1.21.1
google-api-core
google-cloud-dialogflow
django-dialogflow
dialogflow==1.1.1

